refactor: extract reusable helpers from features into @utils
